requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
build==1.2.1
cachetools==5.3.3
cairocffi==1.7.0
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.8.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
fonttools==4.52.4
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
Glances==4.0.7
gradio==4.31.5
gradio_client==0.16.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
imageio==2.34.1
importlib_resources==6.4.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
networkx==3.3
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pandas==2.2.2
Pillow==9.5.0
pip-tools==7.4.1
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
py3nvml==0.2.7
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
pythainlp==5.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.6
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.1
semantic-version==2.10.0
sentry-sdk==2.3.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.35.0
tenacity==8.3.0
tensorboardX==2.6.2.2
tifffile==2024.5.22
timm==1.0.3
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.4
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
wandb==0.17.0
watchdog==4.0.1
watchfiles==0.22.0
webencodings==0.5.1
websockets==11.0.3
xmltodict==0.13.0
